requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
Pillow>=9.0
tqdm>=4.65
pyyaml>=6.0
scikit-learn>=1.3
pandas>=2.0
tensorboard>=2.13
lpips>=0.1.4
